Chore: Factor DataRegistry integration into separate file
