+++ src/index.ts
@@ -1,25 +1,12 @@
 import { JupyterFrontEndPlugin, JupyterFrontEnd, ILayoutRestorer } from "@jupyterlab/application";
 import { IWidgetTracker, WidgetTracker } from "@jupyterlab/apputils";
-import { IRegistry, widgetDataType, CSVDataType } from "@jupyterlab/dataregistry-extension";
+import { IRegistry } from "@jupyterlab/dataregistry-extension";
 import { IDocumentWidget } from "@jupyterlab/docregistry";
 import { Token } from "@phosphor/coreutils";
 import { SpreadsheetModelFactory, JupyterSpreadsheetModel } from "./modelfactory";
 import { SpreadsheetWidget } from "./widget";
 import { SpreadsheetWidgetFactory } from "./widgetfactory";
-import {
-    createConverter,
-    resolveExtensionConverter,
-    relativeNestedDataType,
-    DataTypeStringArg,
-    URLDataType,
-    DataTypeNoArgs,
-    resolveDataType,
-    URLTemplate,
-} from "@jupyterlab/dataregistry";
-import { SpreadsheetModel } from "./model";
-import { defer, BehaviorSubject, merge, Observable, throwError } from "rxjs";
-import { fromFetch } from "rxjs/fetch";
-import { map, distinct, switchMap } from "rxjs/operators";
+import { registerConverters, XLSX_MIMETYPE, XLS_MIMETYPE } from "./registry";
 
 export const ISpreadsheetTracker = new Token("jupyterlab-spreadsheet:tracker");
 export type ISpreadsheetTracker = IWidgetTracker<
@@ -61,145 +48,15 @@ function activateSpreadsheet(
             ".xlsx"
         ],
         mimeTypes: [
-            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-            "application/vnd.ms-excel",
+            XLS_MIMETYPE,
+            XLSX_MIMETYPE,
             "application/octet-stream",
             "text/plain"
         ]
     });
     
     if (registry) {
-        const XLS_MIMETYPE = "application/vnd.ms-excel";
-        const XLSX_MIMETYPE = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
-        const CWF_MIMETYPE = "application/vnd.sheetjs.cwf+json";
-        const SheetJsWorkbookDataType = new DataTypeNoArgs<SpreadsheetModel>(CWF_MIMETYPE);
-
-        // A datatype with an explicit sheet reference.
-        // NOTE: This is _not_ the same as an activeSheet.
-        const SheetJsDataType = new DataTypeStringArg<SpreadsheetModel>(CWF_MIMETYPE, "worksheet");
-        
-        // A datatype for shuttling Base64 around, much like the textDataType
-        const Base64DataType = new DataTypeStringArg<Observable<string>>(
-            "application/octet-stream",
-            "mimetype"
-        ); 
-
-        registry.addConverter(
-            resolveExtensionConverter(".xls", XLS_MIMETYPE),
-            resolveExtensionConverter(".xlsx", XLSX_MIMETYPE),
-            // A converter to go between URLs and Base64 data
-            createConverter({
-                from: URLDataType,
-                to: Base64DataType,
-            }, ({data, type}) => ({
-                type: type,
-                data: data.pipe(
-                    distinct(),
-                    switchMap(i => fromFetch(i)),
-                    switchMap(i => {
-                        if (!i.ok) {
-                            return throwError(new Error("Bad response:" + i));
-                        }
-                        return i.blob()
-                            .then(blob => new Promise((res, rej) => {
-                                // don't you love how most DOM APIs are
-                                // constructed entirely out of duct tape?
-                                const reader = new FileReader();
-                                reader.onerror = rej;
-                                reader.onload = () => {
-                                    const result = reader.result;
-                                    if (result == null || reader instanceof ArrayBuffer) {
-                                        return rej("Failed to load data")
-                                    }
-                                    res(("" + result).replace(/^data:[^;]+;base64,/, ""));
-                                };
-                                reader.readAsDataURL(blob);
-                            }));
-                    })
-                )
-            })),
-            createConverter({
-                from: Base64DataType,
-                to: SheetJsWorkbookDataType,
-            }, ({type, data}) => {
-                if (!(type === XLS_MIMETYPE || type === XLSX_MIMETYPE)) {
-                    return null;
-                }
-                
-                return {
-                    type: void 0,
-                    data: new SpreadsheetModel({ value: data })
-                };
-            }),
-            createConverter({
-                from: SheetJsWorkbookDataType,
-                to: widgetDataType,
-            }, ({ data }) => {
-                return {
-                    type: "Spreadsheet",
-                    data: () => new SpreadsheetWidget({model: data})
-                }
-            }),
-            createConverter({
-                from: SheetJsWorkbookDataType,
-                to: relativeNestedDataType,
-            }, ({data}) => ({
-                    type: void 0,
-                    data: defer(() => {
-                        return merge(
-                            new BehaviorSubject(void 0),
-                            data.workbookChanged
-                        ).pipe(
-                            map(() => data.getSheetNames()
-                                .map(i => "#/sheet/" + i)
-                            )
-                        )
-                    })
-                })
-            ),
-            createConverter({
-                from: resolveDataType,
-                to: SheetJsDataType,
-            }, ({ url }) => {
-                const template = new URLTemplate("/sheet/{sheetName}/", {
-                    "sheetName": URLTemplate.string
-                });
-                const result = template.parse(url.href);
-                if (            	
-                    url.protocol !== "file:" ||	
-                    !(/\.xlsx?$/.test(url.pathname)) ||
-                    result == null
-                ) {
-                    return null;
-                }
-
-                const sheetName = result.sheetName;
-                url.hash = "";
-                const data = SheetJsWorkbookDataType
-                    .getDataset(registry.getURL("" + url));
-                if (data == null) return null;
-                return {
-                    type: sheetName,
-                    data
-                };
-            }),
-            createConverter({
-                from: SheetJsDataType,
-                to: CSVDataType
-            }, ({ data, type }) => {
-                return {
-                    type: void 0,
-                    data: defer(() => {
-                        return merge(
-                            new BehaviorSubject(void 0),
-                            data.workbookChanged
-                        ).pipe(
-                            map(() => data.toCsv(type))
-                        )
-                    })
-                };
-            })
-        );
+        registerConverters(registry)
     }
 
     factory.widgetCreated.connect((sender, widget) => {
